Refactor: Simplify plugin and remove settings
Removes settings and executor, simplifying the plugin to a basic code block processor.
